Split runtime internal modules
